move example to own file
